@@ -1,91 +1,94 @@
 {
-  "name": "web-server",
+  "name": "npm-gui",
   "version": "1.0.0",
-  "description": "npm-gui web-server",
+  "description": "npm-gui",
   "main": "index.js",
   "scripts": {
-    "dev": "node node_modules/concurrently/src/main.js \"npm run dev:server\" \"npm run dev:web-client\" ",
-    "dev:server": "node node_modules/concurrently/src/main.js \"npm run dev:server:webpack\" \"npm run dev:server:nodemon\" ",
-    "dev:server:webpack": "node node_modules/webpack/bin/webpack --config webpack.server.config.js --watch",
-    "dev:server:nodemon": "node node_modules/nodemon/bin/nodemon --watch dist bin/npm-gui 0.0.0.0:9002",
-    "dev:web-client": "node node_modules/webpack-dev-server/bin/webpack-dev-server --config webpack.web-client.config.js --hot",
-    "build": "node node_modules/webpack/bin/webpack -p",
+    "dev": "concurrently \"npm run dev:server\" \"npm run dev:client\" ",
+    "dev:server": "concurrently \"npm run dev:server:webpack\" \"npm run dev:server:nodemon\" ",
+    "dev:server:webpack": "webpack --config webpack.server.config.js --watch --mode=development",
+    "dev:server:nodemon": "nodemon --watch dist bin/npm-gui 0.0.0.0:9002",
+    "dev:client": "webpack-dev-server --config webpack.client.config.js --hot --mode=development",
+    "build": "webpack -p",
     "test:unit": "karma start test/unit/karma.conf.js --single-run",
     "lint": "npm run lint:css && npm run lint:js",
     "lint:css": "node node_modules/stylelint/dist/cli.js 'components/*/**.vue' *.scss --syntax scss",
-    "lint:js": "node node_modules/eslint/bin/eslint \"web-client/components/**/**.vue\" \"web-client/components/**/**.js\" \"server/**/**.js\" ",
+    "lint:js": "node node_modules/eslint/bin/eslint \"client/components/**/**.vue\" \"client/components/**/**.js\" \"server/**/**.js\" ",
     "e2e": "node test/e2e/runner.js",
     "test": "npm run lint:js",
     "test:unit:core": "node node_modules/mocha/bin/mocha ./service/**/*.test.js --compilers js:babel-register --require babel-polyfill"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/npm-gui-base/web-client.git"
+    "url": "git+https://github.com/npm-gui-base/client.git"
   },
   "author": "",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/npm-gui-base/web-client/issues"
+    "url": "https://github.com/npm-gui-base/client/issues"
   },
-  "homepage": "https://github.com/npm-gui-base/web-client#readme",
+  "homepage": "https://github.com/npm-gui-base/client#readme",
   "devDependencies": {
+    "@babel/core": "^7.1.2",
+    "@babel/plugin-transform-async-to-generator": "^7.1.0",
+    "@babel/plugin-transform-runtime": "^7.1.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "axios": "^0.18.0",
     "babel-core": "^6.25.0",
     "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.1",
+    "babel-loader": "^8.0.4",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.23.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
     "babel-register": "^6.24.1",
     "concurrently": "^3.5.0",
-    "css-loader": "^0.28.4",
+    "css-loader": "^1.0.0",
+    "debounce": "^1.2.0",
     "eslint": "^3.19.0",
     "eslint-config-airbnb-base": "^5.0.3",
     "eslint-loader": "^1.9.0",
     "eslint-plugin-import": "^1.16.0",
-    "file-loader": "^0.11.2",
-    "html-webpack-plugin": "^2.29.0",
+    "file-loader": "^2.0.0",
+    "html-webpack-plugin": "^3.2.0",
+    "is-utf8": "^0.2.1",
     "mocha": "^2.5.3",
-    "node-sass": "^4.5.3",
     "nodemon": "^1.11.0",
+    "normalize.css": "^8.0.0",
     "proxyquire": "^1.8.0",
     "should": "^11.2.1",
     "should-sinon": "0.0.5",
     "sinon": "^1.17.7",
     "sinon-stub-promise": "^2.1.0",
-    "url-loader": "^0.5.9",
-    "vue-loader": "^9.9.5",
+    "style-loader": "^0.23.1",
+    "url-loader": "^1.1.2",
+    "vue": "^2.5.17",
+    "vue-loader": "^15.4.2",
+    "vue-router": "^3.0.1",
     "vue-template-compiler": "^2.0.8",
-    "webpack": "^3.2.0",
-    "webpack-dev-server": "^2.5.1",
-    "webpack-node-externals": "^1.6.0"
+    "webpack": "^4.20.2",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.9",
+    "webpack-node-externals": "^1.7.2"
   },
   "engines": {
     "node": ">= 0.10.0",
     "npm": ">= 1.0.0"
   },
   "dependencies": {
-    "axios": "^0.16.2",
     "bluebird": "^3.5.0",
     "body-parser": "^1.17.2",
     "cross-spawn": "^5.1.0",
-    "debounce": "^1.0.2",
     "eslint-plugin-html": "^3.1.0",
     "express": "^4.15.3",
     "fs-access": "^1.0.1",
     "left-pad": "^1.1.3",
-    "normalize.css": "^7.0.0",
     "nsp": "^2.6.3",
     "opn": "^5.1.0",
     "request": "^2.81.0",
     "rimraf": "^2.6.1",
     "rx": "^4.1.0",
-    "sass-flex-mixin": "^1.0.3",
-    "sass-loader": "^6.0.6",
-    "scss-loader": "0.0.1",
-    "style-loader": "^0.18.2",
-    "vue": "^2.0.8",
-    "vue-router": "^2.0.3",
     "ws": "^1.1.4"
   }
 }
