@@ -1,10 +1,11 @@
 import express from 'express';
 import globalModulesController from './globalPackages.controller.js';
 
+import { getGlobalDependencies } from '../../actions/dependencies/getGlobalDependencies';
+
 const globalModulesRouter = express.Router(); // eslint-disable-line
 
-
-globalModulesRouter.get('/', globalModulesController.whenGet);
+globalModulesRouter.get('/', getGlobalDependencies);
 globalModulesRouter.put('/', globalModulesController.whenPut);
 globalModulesRouter.delete('/:name', globalModulesController.whenDelete);
 // others
